688: move FilterDuplicates to common steps
