+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/stores/importer/bayern/EhrenamtskarteBayern.kt
@@ -3,12 +3,12 @@ package app.ehrenamtskarte.backend.stores.importer.bayern
 import app.ehrenamtskarte.backend.stores.importer.ImportConfig
 import app.ehrenamtskarte.backend.stores.importer.addStep
 import app.ehrenamtskarte.backend.stores.importer.bayern.steps.DownloadLbe
-import app.ehrenamtskarte.backend.stores.importer.bayern.steps.FilterDuplicates
 import app.ehrenamtskarte.backend.stores.importer.bayern.steps.FilterLbe
 import app.ehrenamtskarte.backend.stores.importer.bayern.steps.MapFromLbe
 import app.ehrenamtskarte.backend.stores.importer.bayern.steps.PostSanitizeFilter
 import app.ehrenamtskarte.backend.stores.importer.bayern.steps.SanitizeGeocode
 import app.ehrenamtskarte.backend.stores.importer.bayern.steps.Store
+import app.ehrenamtskarte.backend.stores.importer.common.steps.FilterDuplicates
 import app.ehrenamtskarte.backend.stores.importer.common.steps.SanitizeAddress
 import app.ehrenamtskarte.backend.stores.importer.pipelines.Pipeline
 import io.ktor.client.HttpClient
